moved types.d.ts to /src/
